% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_isi.R
\name{calc_isi}
\alias{calc_isi}
\title{Calculate insulin-sensitivity index}
\usage{
calc_isi(glu = NULL, ins = NULL, nefa = NULL, bmi = NULL,
  body_weight = NULL, index = NULL)
}
\arguments{
\item{glu}{An ordered vector of glucose values.}

\item{ins}{An ordered vector of insulin values.}

\item{nefa}{An ordered vector of NEFA values.}

\item{bmi}{Body-mass index.}

\item{body_weight}{Body weight in kg.}

\item{index}{Insulin sensitivity index as a string. This can be one of 
"Cederholm", "Gutt" (aka "ISI 0,120"), "Matsuda", "NEFA-ISI", 
"Stumvoll-ISI" or "Stumvoll-MCR".}
}
\value{
An integer.
}
\description{
Calculate insulin-sensitivity index
}
