% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogtt-based.R
\name{cederholm}
\alias{cederholm}
\title{Cederholm insulin sensitivity index}
\usage{
cederholm(bw, g, i)
}
\arguments{
\item{bw}{Body weight(kg).}

\item{g}{Vector with glucose values.}

\item{i}{Vector with insulin values.}
}
\value{
An integer.
}
\description{
Cederholm insulin sensitivity index
}

