% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cederholm_index.R
\name{cederholm_index}
\alias{cederholm_index}
\title{Cederholm index}
\usage{
cederholm_index(glu, ins, body_weight)
}
\arguments{
\item{glu}{An ordered vector of glucose values.}

\item{ins}{An ordered vector of insulin values.}

\item{body_weight}{Body weight in kg.}
}
\value{
A double.
}
\description{
Cederholm index
}
