% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasting-based.R
\name{FIRI}
\alias{FIRI}
\title{FIRI insulin sensitivity index}
\usage{
FIRI(g, i)
}
\arguments{
\item{g}{Fasting glucose.}

\item{i}{Fasting insulin.}
}
\value{
An integer.
}
\description{
FIRI insulin sensitivity index
}

