% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firi_index.R
\name{firi_index}
\alias{firi_index}
\title{FIRI index}
\usage{
firi_index(glu, ins)
}
\arguments{
\item{glu}{The value of fasting glucose.}

\item{ins}{The value of fasting insulin.}
}
\value{

}
\description{
FIRI index
}
