% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gmi.R
\name{Gmi}
\alias{Gmi}
\title{Title}
\usage{
Gmi(
  X,
  y,
  beta,
  penalty = "lasso",
  penalty.factor = rep(1, ncol(X)),
  inter.penalty.factor = 1,
  lam.list,
  lambda.min.ratio = 0.001,
  max.iter = 50,
  n.lambda = 100,
  eps = 1e-15,
  tune = "EBIC",
  ebic.gamma = 1,
  eta = 0.5
)
}
\arguments{
\item{X}{input matrix, of dimension nobs x nvars; each row is an observation}

\item{y}{response}

\item{eta}{}
}
\description{
Title
}
