% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogtt-based.R
\name{gutt}
\alias{gutt}
\title{Gutt insulin sensitivity index}
\usage{
gutt(bw, g, i)
}
\arguments{
\item{bw}{Body weight (kg)}

\item{g}{Vector with glucose values.}

\item{i}{Vector with insulin values.}
}
\value{
An integer.
}
\description{
Gutt insulin sensitivity index
}

