% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutt_index.R
\name{gutt_index}
\alias{gutt_index}
\title{Title}
\usage{
gutt_index(glu, ins, body_weight)
}
\arguments{
\item{glu}{An ordered vector of glucose values.}

\item{ins}{An ordered vector of insulin values.}

\item{body_weight}{Body weight in kg.}
}
\value{
A double.
}
\description{
Title
}
