% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homa_ir_index.R
\name{homa_ir_index}
\alias{homa_ir_index}
\title{HOMA-IR index}
\usage{
homa_ir_index(glu, ins)
}
\arguments{
\item{glu}{The value of fasting glucose.}

\item{ins}{The value of fasting insulin.}
}
\value{

}
\description{
HOMA-IR index
}
