% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasting-based.R
\name{homaIR}
\alias{homaIR}
\title{HOMA-IR insulin sensitivity index}
\usage{
homaIR(g, i)
}
\arguments{
\item{g}{Fasting glucose.}

\item{i}{Fasting insulin.}
}
\value{
An integer.
}
\description{
HOMA-IR insulin sensitivity index
}

