% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intertomain.R
\name{intertomain}
\alias{intertomain}
\title{convert an interaction name list into the corresponding main effect for enforcing      strong hierachy}
\usage{
intertomain(interNameList, p)
}
\arguments{
\item{interNameList}{interaction name list input}

\item{p}{dimensionality}
}
\description{
convert an interaction name list into the corresponding main effect for enforcing      strong hierachy
}
