% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_numeric_or_null.R
\name{is_numeric_or_null}
\alias{is_numeric_or_null}
\title{Numeric or NULL objects}
\usage{
is_numeric_or_null(...)
}
\arguments{
\item{...}{R objects to check.}
}
\value{
a logical.
}
\description{
Numeric or NULL objects
}
