% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasting-based.R
\name{ISI_fasting}
\alias{ISI_fasting}
\title{Calculate insulin sensitivity from fasting values}
\usage{
ISI_fasting(glu = NA, ins = NA, nefa = NA, index = NA)
}
\arguments{
\item{glu}{Fasting glucose.}

\item{ins}{Fasting insulin.}

\item{nefa}{Fasting NEFA.}

\item{index}{Insulin sensitivity index as a string. This can be one of
"FIRI", "HOMA-IR", "QUICKI" or "Revised QUICKI".}
}
\value{
An integer.
}
\description{
Calculate insulin sensitivity from fasting values
}
\examples{
df <- data.frame(glucose = 90, insulin = 57, nefa = 287)
insuSen <- ISI_fasting(glu = df$glucose, ins = df$insulin,
nefa = df$nefa, index = "revised QUICKI")

# Using sample data
require(dplyr)
OGTT_tidy \%>\%
filter(timePoint == -60) \%>\% # keep only fasting values
group_by(ID) \%>\% # calculate index for each subject
summarise(InsuSen = ISI_fasting(glu = glucose_mgdl,
                                ins = insulin_microU, nefa = nefa_umolL,
                                index = "revised QUICKI"))


}
