% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogtt-based.R
\name{ISI_OGTT}
\alias{ISI_OGTT}
\title{Calculate insulin sensitivity from OGTT values}
\usage{
ISI_OGTT(glu = NA, ins = NA, nefa = NA, bmi = NA, bw = NA,
  index = NA)
}
\arguments{
\item{glu}{Glucose.}

\item{ins}{Insulin.}

\item{nefa}{NEFA.}

\item{bmi}{Body-mass index.}

\item{bw}{Body weight in kilos.}

\item{index}{Insulin sensitivity index as a string. This can be one of "Cederholm",
"Gutt" (aka "ISI 0,120"), "Matsuda", "NEFA-ISI", "Stumvoll-ISI" or "Stumvoll-MCR".}
}
\value{
An integer.
}
\description{
Calculate insulin sensitivity from OGTT values
}
