% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikbinomial.R
\name{loglikbinomial}
\alias{loglikbinomial}
\title{Binomial Log-Likelihood Function}
\usage{
loglikbinomial(X, y, beta)
}
\arguments{
\item{X}{A numeric matrix of predictors, with dimensions \code{n x p}.}

\item{y}{A binary response vector of length \code{n}, with values in \code{0} or \code{1}.}

\item{beta}{A numeric coefficient vector of length \code{p}.}
}
\value{
A single numeric value representing the binomial log-likelihood
}
\description{
Binomial Log-Likelihood Function
}
\examples{
set.seed(1)
X <- matrix(rnorm(100), nrow = 20)
beta <- runif(ncol(X))
eta <- X \%*\% beta
prob <- Sigmoid(eta)
y <- rbinom(20, 1, prob)
loglikbinomial(X, y, beta)
}
