% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogtt-based.R
\name{matsuda}
\alias{matsuda}
\title{Matsuda Index.}
\usage{
matsuda(g, i)
}
\arguments{
\item{g}{A vector with ordered glucose values.}

\item{i}{A vector with ordered insulin values.}
}
\value{
An integer.
}
\description{
Calculate the Matsuda insulin sensitivity index.
}
\examples{
glu <- c(4.6, 7.2, 9.8, 6.8, 5.16)
ins <- c(33, 263, 797, 416, 274)
mat_isi <- matsuda(glu, ins)



}

