% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsuda_index.R
\name{matsuda_index}
\alias{matsuda_index}
\title{Matsuda Index (Composite index)}
\usage{
matsuda_index(glu, ins)
}
\arguments{
\item{glu}{An ordered vector of glucose values
(0', 30', 60', 90', 120').}

\item{ins}{An ordered vector of insulin values.
(0', 30', 60', 90', 120').}
}
\value{
a double.
}
\description{
Matsuda Index (Composite index)
}
