% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefa_index.R
\name{nefa_index}
\alias{nefa_index}
\title{NEFA insulin index}
\usage{
nefa_index(ins, bmi, nefa)
}
\arguments{
\item{ins}{An ordered vector of insulin values.}

\item{bmi}{Body-mass index.}

\item{nefa}{An ordered vector of NEFA values.}
}
\value{
A double.
}
\description{
NEFA insulin index
}
