% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogtt-based.R
\name{nefaISI}
\alias{nefaISI}
\title{NEFA-ISI insulin sensitivity index}
\usage{
nefaISI(bmi, i, n)
}
\arguments{
\item{bmi}{BMI.}

\item{i}{Insulin.}

\item{n}{NEFA.}
}
\value{
An integer.
}
\description{
Described by Wagner at al., 2016 (http://goo.gl/cwJNz7)
}

