% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_Gmi.R
\name{predict_Gmi}
\alias{predict_Gmi}
\title{Model prediction based on a fitted Gmi object.}
\usage{
predict_Gmi(
  object,
  newdata = NULL,
  type = c("link", "response", "class"),
  allpath = FALSE
)
}
\arguments{
\item{object}{Fitted \code{'Gmi'} model object.}

\item{newdata}{Matrix of new values for \code{x} at which predictions are to be made, without the intercept term.}

\item{type}{Type of prediction required.
Type \code{'response'} gives the fitted probabilities,
Type \code{'link'} returns the linear predictors for models;
Type \code{'class'} produces the class label corresponding to the maximum probability (0-1 labels).}

\item{allpath}{allpath = T will output all the predictions on the solution path. allpath = FALSE will only output the one the criterion selected in the \code{'Gmi'} object.}
}
\value{
The object returned depends on type.
\itemize{
\item If \code{type = "class"}, returns logical (TRUE/FALSE) values for class labels.
\item If \code{type = "response"}, returns fitted probabilities.
\item If \code{type = "link"}, returns the linear predictors.
}
}
\description{
Similar to the usual predict methods, this function returns predictions from a fitted \code{'Gmi'} object
}
