% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasting-based.R
\name{QUICKI}
\alias{QUICKI}
\title{QUICKI insulin sensitivity index}
\usage{
QUICKI(g, i)
}
\arguments{
\item{g}{Fasting glucose.}

\item{i}{Fasting insulin.}
}
\value{
An integer.
}
\description{
QUICKI insulin sensitivity index
}

