% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicki_index.R
\name{quicki_index}
\alias{quicki_index}
\title{QUICKI index}
\usage{
quicki_index(glu, ins)
}
\arguments{
\item{glu}{The value of fasting glucose.}

\item{ins}{The value of fasting insulin.}
}
\value{
A double.
}
\description{
QUICKI index
}
