% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasting-based.R
\name{QUICKIR}
\alias{QUICKIR}
\title{Revised QUICKI insulin sensitivity index}
\usage{
QUICKIR(g, i, n)
}
\arguments{
\item{g}{Fasting glucose.}

\item{i}{Fasting insulin.}

\item{n}{Fasting NEFA.}
}
\value{
An integer.
}
\description{
Source: http://goo.gl/OJ5OZi
}

