% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revised_quicki_index.R
\name{revised_quicki_index}
\alias{revised_quicki_index}
\title{Revised QUICKI index}
\usage{
revised_quicki_index(glu, ins, nefa)
}
\arguments{
\item{glu}{The value of fasting glucose.}

\item{ins}{The value of fasting insulin.}

\item{nefa}{The value of fasting NEFA.}
}
\value{

}
\description{
Revised QUICKI index
}
