% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigmoid.R
\name{Sigmoid}
\alias{Sigmoid}
\title{Sigmoid Activation Function}
\usage{
Sigmoid(z)
}
\arguments{
\item{z}{numeric vector}
}
\value{
numeric vector or matrix with the sigmoid transformation applied.
}
\description{
Sigmoid Activation Function
}
\examples{
sigmoid(0)
}
