% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stumvoll_isi_index.R
\name{stumvoll_isi_index}
\alias{stumvoll_isi_index}
\title{Stumvoll ISI index}
\usage{
stumvoll_isi_index(glu, ins, bmi)
}
\arguments{
\item{glu}{An ordered vector of glucose values
(0', 30', 60', 90', 120').}

\item{ins}{An ordered vector of insulin values.
(0', 30', 60', 90', 120').}

\item{bmi}{Body-mass index.}
}
\value{

}
\description{
Stumvoll ISI index
}
