% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stumvoll_mcr_index.R
\name{stumvoll_mcr_index}
\alias{stumvoll_mcr_index}
\title{Stumvoll MCR index}
\usage{
stumvoll_mcr_index(glu, ins, bmi)
}
\arguments{
\item{glu}{An ordered vector of glucose values
(0', 30', 60', 90', 120').}

\item{ins}{An ordered vector of insulin values.
(0', 30', 60', 90', 120').}

\item{bmi}{Body-mass index.}
}
\value{
A double.
}
\description{
Stumvoll MCR index
}
