% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogtt-based.R
\name{stumvollisi}
\alias{stumvollisi}
\title{Stumvoll Insulin Sensitivity Index}
\usage{
stumvollisi(bmi, g, i)
}
\arguments{
\item{bmi}{Body-mass index.}

\item{g}{Glucose at 90min.}

\item{i}{Insulin at 120min.}
}
\value{
An integer.
}
\description{
Stumvoll Insulin Sensitivity Index
}

