% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogtt-based.R
\name{stumvollmcr}
\alias{stumvollmcr}
\title{Stumvoll Metabolic Clearance Rate}
\usage{
stumvollmcr(bmi, i, g)
}
\arguments{
\item{bmi}{Body-mass Index}

\item{i}{Insulin at 120min}

\item{g}{Glucose at 90min}
}
\value{
An integer.
}
\description{
Stumvoll Metabolic Clearance Rate
}

