% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OGTT_tidy}
\alias{OGTT_tidy}
\title{Sample tidy data from an OGTT in tidy.}
\format{A data frame with 60 rows and 9 variables:
\describe{
  \item{ID}{subject identifier}
  \item{weight_kg}{subject's weight, in kilograms}
  \item{bmi}{body-mass index, in kg/m^2}
  \item{timePoint}{minutes after the OGTT start; -60 is fasting}
  \item{glucose_mgl}{glucose values in mg/dl}
  \item{insulin_pmoll}{insulin values in pmol/L (SI)}
  \item{nefa_umolL}{NEFA values in µmol/L}
  \item{glucose_mmoll}{glucose values in mmol/L (SI)}
  \item{insulin_microU}{insulin values in µU/L}
}}
\usage{
OGTT_tidy
}
\description{
A dataset containing fasting and 2-hour OGTT values.
}
\keyword{datasets}
