% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_Gmi.R
\name{predict_Gmi}
\alias{predict_Gmi}
\title{Title}
\usage{
predict_Gmi(
  object,
  newdata = NULL,
  type = c("link", "response", "class"),
  allpath = FALSE
)
}
\arguments{
\item{allpath}{}
}
\description{
Title
}
