% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBfusedlasso.R
\name{SBfusedlasso}
\alias{SBfusedlasso}
\title{Title}
\usage{
SBfusedlasso(
  X,
  y,
  a_0,
  beta,
  lam,
  eta,
  rho1 = 1,
  rho2 = 1,
  penalty.type,
  epsilon1 = 4e-07,
  epsilon2 = 4e-07,
  maxiter1 = 100,
  maxiter2 = 100,
  pf
)
}
\arguments{
\item{X}{input matrix, of dimension nobs x nvars; each row is an observation}

\item{y}{response variable, of dimension nobs x}

\item{pf}{}
}
\description{
Title
}
