% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigmoid.R
\name{Sigmoid}
\alias{Sigmoid}
\title{Sigmoid Activation Function}
\usage{
Sigmoid(z)
}
\arguments{
\item{z}{A numeric vector.}
}
\value{
A numeric vector with the sigmoid transformation applied.
}
\description{
Sigmoid Activation Function
}
\examples{
Sigmoid(0)
}
